+++ src/main/java/subway/presentation/SubwayExceptionHandler.java
@@ -1,0 +1,27 @@
+package subway.presentation;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+import subway.exception.GlobalException;
+import subway.presentation.dto.response.BadResponse;
+
+@RestControllerAdvice
+public class SubwayExceptionHandler {
+
+    @ExceptionHandler
+    public ResponseEntity<BadResponse> handleGlobalException(GlobalException e) {
+        return ResponseEntity
+                .status(HttpStatus.BAD_REQUEST)
+                .body(new BadResponse(e.getMessage()));
+    }
+
+    @ExceptionHandler
+    public ResponseEntity<BadResponse> handleException(Exception e) {
+        e.printStackTrace();
+        return ResponseEntity
+                .status(HttpStatus.INTERNAL_SERVER_ERROR)
+                .body(new BadResponse("예기치 못한 오류가 발생하였습니다."));
+    }
+}
