+++ src/main/java/subway/presentation/dto/response/BadResponse.java
@@ -1,0 +1,13 @@
+package subway.presentation.dto.response;
+
+public class BadResponse {
+    private final String errorMessage;
+
+    public BadResponse(String errorMessage) {
+        this.errorMessage = errorMessage;
+    }
+
+    public String getErrorMessage() {
+        return errorMessage;
+    }
+}
